Добавлена проверка наличия прав на удаление команды при выполнении метода execute в DeleteTeamAction.

+++ src/org/cherchgk/actions/team/DeleteTeamAction.java
@@ -3,6 +3,7 @@ package org.cherchgk.actions.team;
 import com.opensymphony.xwork2.Action;
 import com.opensymphony.xwork2.ActionSupport;
 import org.cherchgk.domain.Team;
+import org.cherchgk.security.PermissionChecker;
 import org.cherchgk.services.TeamService;
 import org.cherchgk.utils.ActionContextHelper;
 
@@ -21,6 +22,7 @@ public class DeleteTeamAction extends ActionSupport {
     @Override
     public String execute() throws Exception {
         String teamId = ActionContextHelper.getRequestParameterValue("teamId");
+        PermissionChecker.checkPermissions("team:delete:" + teamId);
         Team team = teamService.find(Long.valueOf(teamId));
         tournamentId = team.getTournament().getId();
         teamService.delete(team);
